+++ src/content-scripts/components/GenerateAIDescription/DescriptionGeneratorButton.ts
@@ -16,33 +16,46 @@ export const DescriptionGeneratorButton = () => {
     </span>
     <tool-tip for="ai-description-gen">Generate PR description</tool-tip>`,
         onclick: handleSubmit,
-
     });
 
     return descriptionGeneratorButton;
 };
 
 const handleSubmit = async () => {
-    const logo = document.getElementById("ai-description-button-logo") ?? null;
+    const logo = document.getElementById("ai-description-button-logo");
+    const button = document.getElementById("ai-description-button");
 
     try {
         if (!(await isLoggedIn())) {
             return window.open(SUPABASE_LOGIN_URL, "_blank");
         }
 
-        if (!logo) {
+        if (!logo || !button) {
             return;
         }
-        logo.classList.toggle("animate-spin");
-        const descriptionStream = await getAiDescription();
+
+        const descriptionConfig = await getAIDescriptionConfig();
+
+        if (!descriptionConfig) {
+            return;
+        }
 
         logo.classList.toggle("animate-spin");
+        button.classList.toggle("pointer-events-none");
+
+
+        const { protocol, hostname, pathname } = window.location;
+        const descriptionStream = await getAiDescription(`${protocol}//${hostname}${pathname}`);
+
+        logo.classList.toggle("animate-spin");
+        button.classList.toggle("pointer-events-none");
 
         const textArea = document.getElementsByName(GITHUB_PR_COMMENT_TEXT_AREA_SELECTOR)[0] as HTMLTextAreaElement;
 
         insertTextAtCursor(textArea, descriptionStream);
     } catch (error: unknown) {
         logo?.classList.toggle("animate-spin");
+        button?.classList.toggle("pointer-events-none");
 
         if (error instanceof Error) {
             alert(error.message);
@@ -51,19 +64,16 @@ const handleSubmit = async () => {
     }
 };
 
-export const getAiDescription = async () => {
-    const url = getPullRequestAPIURL(window.location.href);
+export const getAiDescription = async (prUrl: string) => {
+    const prApiUrl = await getPullRequestAPIURL(prUrl);
+
     const descriptionConfig = await getAIDescriptionConfig();
 
     if (!descriptionConfig) {
         throw new Error("Configuration file is empty!");
     }
 
-    if (!descriptionConfig.enabled) {
-        throw new Error("AI PR description is disabled!");
-    }
-
-    const [diff, commitMessages] = await getDescriptionContext(url, descriptionConfig.config.source);
+    const [diff, commitMessages] = await getDescriptionContext(prApiUrl, descriptionConfig.config.source);
 
     if (!diff && !commitMessages) {
         throw new Error(`No input context was generated.`);
@@ -72,7 +82,7 @@ export const getAiDescription = async () => {
         throw new Error(`Max input length exceeded. Try setting the description source to commit-messages.`);
     }
     const token = await getAuthToken();
-    const descriptionStream = await generateDescription(
+    const description = await generateDescription(
         token,
         descriptionConfig.config.language,
         descriptionConfig.config.length,
@@ -82,10 +92,11 @@ export const getAiDescription = async () => {
         commitMessages,
     );
 
-    if (!descriptionStream) {
+    if (!description) {
         throw new Error("No description was generated!");
     }
 
-    return descriptionStream;
+    return description;
 };
+