+++ src/popup/pages/home.tsx
@@ -6,39 +6,38 @@ import {
     HiPencil,
     HiUserCircle,
 } from "react-icons/hi2";
-import { useEffect, useState } from "react";
-import OpenSaucedLogo from "../assets/opensauced-logo.svg";
-import { useAuth } from "../hooks/useAuth";
-import { useOpensaucedUserCheck } from "../hooks/useOpensaucedUserCheck";
+import { FiSettings } from "react-icons/fi";
+import OpenSaucedLogo from "../../assets/opensauced-logo.svg";
+import { useAuth } from "../../hooks/useAuth";
+import { useOpensaucedUserCheck } from "../../hooks/useOpensaucedUserCheck";
 import { Profile } from "./profile";
 import { goTo } from "react-chrome-extension-router";
 import AIPRDescription from "./aiprdescription";
 import PostOnHighlight from "./posthighlight";
-import { getHighlights } from "../utils/fetchOpenSaucedApiData";
+import { getHighlights } from "../../utils/fetchOpenSaucedApiData";
 
 import Help from "./help";
-import { OPEN_SAUCED_INSIGHTS_DOMAIN } from "../constants";
-interface Highlight {
-    highlight: string;
-    title: string;
-    name: string;
-    url: string;
-    login: string;
-}
-
+import { useEffect, useState } from "react";
+import Settings from "./settings";
+import { OPEN_SAUCED_INSIGHTS_DOMAIN } from "../../constants";
+import type { Highlight } from "../../ts/types";
+import { useIsGithubPRPageCheck } from "../../hooks/useGithubPRPageCheck";
 
 const Home = () => {
     const { user } = useAuth();
     const { currentTabIsOpensaucedUser, checkedUser } = useOpensaucedUserCheck();
+    const { isGithubPRPage, prUrl, prTitle } = useIsGithubPRPageCheck();
     const [highlights, setHighlights] = useState<Highlight[]>([]);
     const [currentPage, setCurrentPage] = useState(0);
-    const [currentName, setCurrentName] = useState<string>("");
-
 
     useEffect(() => {
         const fetchHighlights = async () => {
             try {
                 const userHighlightsData = await getHighlights();
+
+                if (!userHighlightsData) {
+                    return;
+                }
                 const highlights = userHighlightsData.data;
 
                 setHighlights(highlights);
@@ -57,13 +56,6 @@ const Home = () => {
     const handleNext = () => {
         setCurrentPage(prevPage => prevPage + 1);
     };
-
-    useEffect(() => {
-        // Update the current name when the highlight changes
-        if (highlights[currentPage]?.login) {
-            setCurrentName(highlights[currentPage].login);
-        }
-    }, [highlights, currentPage]);
 
     return (
         <div className="p-4 bg-slate-800">
@@ -185,15 +177,16 @@ const Home = () => {
                             AI Configuration
                         </button>
 
-                        <button
-                            className="flex items-center bg-slate-700 hover:bg-slate-700/70 hover:text-orange text-white gap-2 p-1.5 px-3 w-full rounded-sm font-medium text-sm"
-                            onClick={() => {
-                                goTo(PostOnHighlight);
-                            }}
-                        >
-                            <HiPencil />
-              Post Highlight
-                        </button>
+                        {isGithubPRPage && (
+                            <button
+                                className="flex items-center bg-slate-700 hover:bg-slate-700/70 hover:text-orange text-white gap-2 p-1.5 px-3 w-full rounded-sm font-medium text-sm"
+                                onClick={() => {
+                                    goTo(PostOnHighlight, { prUrl, prTitle });
+                                }}
+                            >
+                                <HiPencil />
+                                Post Highlight
+                            </button>)}
 
                         {currentTabIsOpensaucedUser && (
                             <button
@@ -223,6 +216,16 @@ const Home = () => {
                     >
                         <HiOutlineQuestionMarkCircle />
                         Help
+                    </button>
+
+                    <button
+                        className="flex items-center bg-slate-700 hover:bg-slate-700/70 hover:text-orange text-white gap-2 p-1.5 px-3 w-fit rounded-sm font-medium text-sm"
+                        onClick={() => {
+                            goTo(Settings);
+                        }}
+                    >
+                        <FiSettings />
+            Settings
                     </button>
                 </footer>
             </div>